service: Implement counting of available services
[service] Implement counting of available services. JB#56225 OMP#JOLLA-552

This adds a array of hash tables for each service to be used for
counting how many different types of available services service.c has.
This requires no re-counting of the services every time the amount is
requested. Changes to counters are done when the services are created
and have available network and when the network is removed.