rtln: Extend notify and better support IPv6 routes set via RA
[rtnl] Extend notify and better support IPv6 routes set via RA. JB#58105

Extend the notify to support IPv6 better when gateways are added and
removed. These use the destination as well as metric value (TODO: get
metric from struct in6_rtmsg) for more complex use.

Add toggle to pass also the routes set by RA (protocol type RTPROT_RA)
to the functions that do request it. For this an additional field to the
notify and a function to enable RTPROT_RA support was added.

In some cases it is also useful to have route request available for the
plugins if there is a plugin monitoring a more complex network
configuration where, for example, additional not-managed interfaces
(blacklisted) are getting a default route assigned by kernel via route
advertisements and should ensure that the route is not used by default
when ConnMan assigned route is to be preferred.