[ofono] Register to rtnl for route changes and enforce default route. JB#58105
This implements a rtnl listener with the new notifier functions to
monitor for route changes and to request all routes after the notifier
has been registered. It also sets to handle the RA enabled routes as
well in order to fully control the mobile data routes.

The other default routes are simply removed when mobile data connection
is established and any new default route added for other than the
current mobile data index are being removed. This is managed with a
simple singly linked list. If mobile data is not connected all new
default routes are added to that list and it is cleaned when connection
is established. Only interfaces that have "rmnet_data" prefix are
managed in this sense.